Adding handlers for open and close of film popup
